Integrate ALIKED + LightGlue into stitching_detailed sample
  Add DNN-based feature extraction and matching as new options to the
  existing stitching_detailed.cpp sample, making it accessible to users
  who already use this tool.

  New command-line options:
  - --features aliked: Use ALIKED for feature extraction
  - --matcher lightglue: Use LightGlue for feature matching
  - --aliked_model <path>: Path to ALIKED ONNX model
  - --lightglue_model <path>: Path to LightGlue ONNX model
  - --lg_score_thresh <float>: LightGlue confidence threshold

  Usage:
    ./stitching_detailed img1.jpg img2.jpg --features aliked --matcher lightglue \
      --aliked_model aliked.onnx --lightglue_model lightglue.onnx
